refactor(test-harness): hoist Comparison and Role to the crate root
Both are used across more than one subsystem: Comparison by the output writers
and by Benchmark::write_reports, Role by ToolchainMatrix and the summary stats.
Their lowest common parent is the crate root, so each becomes a root module.
