+++ solx-benchmark-converter/src/output/summary/suite_stats.rs
@@ -16,11 +16,11 @@ use crate::output::summary::diff_counter::DiffCounter;
 use crate::output::summary::failure_regression::FailureRegression;
 use crate::output::summary::failure_regressions::FailureRegressions;
 use crate::output::summary::paired_bytes::PairedBytes;
-use crate::output::summary::role::Role;
 use crate::output::summary::suite_failures::SuiteFailures;
 use crate::output::summary::suite_row::SuiteRow;
 use crate::output::summary::toolchain_matrix::ToolchainMatrix;
 use crate::output::summary::top_movers::TopMovers;
+use crate::role::Role;
 use crate::utils::commas;
 use crate::utils::median;
 use crate::utils::relative_percent;
