+++ solx-benchmark-converter/src/role.rs
